+++ src/main/java/org/folio/linked/data/integration/kafka/sender/search/WorkCreateMessageSender.java
@@ -9,6 +9,7 @@ import static org.folio.ld.dictionary.ResourceTypeDictionary.WORK;
 import static org.folio.linked.data.domain.dto.ResourceIndexEventType.CREATE;
 import static org.folio.linked.data.util.Constants.STANDALONE_PROFILE;
 import static org.folio.linked.data.util.ResourceUtils.extractWorkFromInstance;
+import static org.folio.linked.data.util.ResourceUtils.isNonLightResourceOfType;
 
 import java.util.Collection;
 import java.util.List;
@@ -40,10 +41,10 @@ public class WorkCreateMessageSender implements CreateMessageSender {
 
   @Override
   public Collection<Resource> apply(Resource resource) {
-    if (resource.isOfType(WORK)) {
+    if (isNonLightResourceOfType(resource, WORK)) {
       return singletonList(resource);
     }
-    if (resource.isOfType(INSTANCE)) {
+    if (isNonLightResourceOfType(resource, INSTANCE)) {
       triggerParentWorkUpdate(resource);
     }
     return emptyList();